+++ dashboard/routes/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', async (req, res) => {
+    res.render('index');
+});
+
+module.exports = router;
